+++ commands/cloudapi-v6/waiter/state.go
@@ -8,8 +8,11 @@ import (
 )
 
 func ServerStateInterrogator(c *core.CommandConfig, objId string) (*string, error) {
-	obj, _, err := c.CloudApiV6Services.Servers().Get(viper.GetString(core.GetFlagName(c.NS, cloudapiv6.ArgDataCenterId)), objId)
+	obj, resp, err := c.CloudApiV6Services.Servers().Get(viper.GetString(core.GetFlagName(c.NS, cloudapiv6.ArgDataCenterId)), objId)
 	if err != nil {
+		if resp != nil && resp.HttpNotFound() {
+			return nil, nil // let it wait longer, maybe the resource is being created
+		}
 		return nil, err
 	}
 	if metadata, ok := obj.GetMetadataOk(); ok && metadata != nil {
@@ -21,8 +24,11 @@ func ServerStateInterrogator(c *core.CommandConfig, objId string) (*string, erro
 }
 
 func K8sClusterStateInterrogator(c *core.CommandConfig, objId string) (*string, error) {
-	obj, _, err := c.CloudApiV6Services.K8s().GetCluster(objId)
+	obj, resp, err := c.CloudApiV6Services.K8s().GetCluster(objId)
 	if err != nil {
+		if resp != nil && resp.HttpNotFound() {
+			return nil, nil // let it wait longer, maybe the resource is being created
+		}
 		return nil, err
 	}
 	if metadata, ok := obj.GetMetadataOk(); ok && metadata != nil {
@@ -34,9 +40,12 @@ func K8sClusterStateInterrogator(c *core.CommandConfig, objId string) (*string, 
 }
 
 func K8sNodeStateInterrogator(c *core.CommandConfig, objId string) (*string, error) {
-	obj, _, err := c.CloudApiV6Services.K8s().GetNode(viper.GetString(core.GetFlagName(c.NS, constants.FlagClusterId)),
+	obj, resp, err := c.CloudApiV6Services.K8s().GetNode(viper.GetString(core.GetFlagName(c.NS, constants.FlagClusterId)),
 		viper.GetString(core.GetFlagName(c.NS, constants.FlagNodepoolId)), objId)
 	if err != nil {
+		if resp != nil && resp.HttpNotFound() {
+			return nil, nil // let it wait longer, maybe the resource is being created
+		}
 		return nil, err
 	}
 	if metadata, ok := obj.GetMetadataOk(); ok && metadata != nil {
@@ -48,8 +57,11 @@ func K8sNodeStateInterrogator(c *core.CommandConfig, objId string) (*string, err
 }
 
 func K8sNodePoolStateInterrogator(c *core.CommandConfig, objId string) (*string, error) {
-	obj, _, err := c.CloudApiV6Services.K8s().GetNodePool(viper.GetString(core.GetFlagName(c.NS, constants.FlagClusterId)), objId)
+	obj, resp, err := c.CloudApiV6Services.K8s().GetNodePool(viper.GetString(core.GetFlagName(c.NS, constants.FlagClusterId)), objId)
 	if err != nil {
+		if resp != nil && resp.HttpNotFound() {
+			return nil, nil // let it wait longer, maybe the resource is being created
+		}
 		return nil, err
 	}
 	if metadata, ok := obj.GetMetadataOk(); ok && metadata != nil {
@@ -61,8 +73,11 @@ func K8sNodePoolStateInterrogator(c *core.CommandConfig, objId string) (*string,
 }
 
 func NatGatewayStateInterrogator(c *core.CommandConfig, objId string) (*string, error) {
-	obj, _, err := c.CloudApiV6Services.NatGateways().Get(viper.GetString(core.GetFlagName(c.NS, cloudapiv6.ArgDataCenterId)), objId)
+	obj, resp, err := c.CloudApiV6Services.NatGateways().Get(viper.GetString(core.GetFlagName(c.NS, cloudapiv6.ArgDataCenterId)), objId)
 	if err != nil {
+		if resp != nil && resp.HttpNotFound() {
+			return nil, nil // let it wait longer, maybe the resource is being created
+		}
 		return nil, err
 	}
 	if metadata, ok := obj.GetMetadataOk(); ok && metadata != nil {
@@ -74,8 +89,11 @@ func NatGatewayStateInterrogator(c *core.CommandConfig, objId string) (*string, 
 }
 
 func NetworkLoadBalancerStateInterrogator(c *core.CommandConfig, objId string) (*string, error) {
-	obj, _, err := c.CloudApiV6Services.NetworkLoadBalancers().Get(viper.GetString(core.GetFlagName(c.NS, cloudapiv6.ArgDataCenterId)), objId)
+	obj, resp, err := c.CloudApiV6Services.NetworkLoadBalancers().Get(viper.GetString(core.GetFlagName(c.NS, cloudapiv6.ArgDataCenterId)), objId)
 	if err != nil {
+		if resp != nil && resp.HttpNotFound() {
+			return nil, nil // let it wait longer, maybe the resource is being created
+		}
 		return nil, err
 	}
 	if metadata, ok := obj.GetMetadataOk(); ok && metadata != nil {
@@ -87,8 +105,11 @@ func NetworkLoadBalancerStateInterrogator(c *core.CommandConfig, objId string) (
 }
 
 func ApplicationLoadBalancerStateInterrogator(c *core.CommandConfig, objId string) (*string, error) {
-	obj, _, err := c.CloudApiV6Services.ApplicationLoadBalancers().Get(viper.GetString(core.GetFlagName(c.NS, cloudapiv6.ArgDataCenterId)), objId)
+	obj, resp, err := c.CloudApiV6Services.ApplicationLoadBalancers().Get(viper.GetString(core.GetFlagName(c.NS, cloudapiv6.ArgDataCenterId)), objId)
 	if err != nil {
+		if resp != nil && resp.HttpNotFound() {
+			return nil, nil // let it wait longer, maybe the resource is being created
+		}
 		return nil, err
 	}
 	if metadata, ok := obj.GetMetadataOk(); ok && metadata != nil {