+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 - Added more descriptive error/help messages for required flags when --type is used in the 'ionosctl server create' command.
+- Added '--promote-volume' (which requires '--wait-for-state') to 'ionosctl server create --type CUBE' to automatically promote an attached DAS as the server's bootVolume after creation.
+
+### Fixed
+- Fixed --wait-for-state instantly erroring out with a 404 for certain resources, instead of waiting for the resource to appear.
 - Add support for Kafka Users
 - Add Kafka version tab-completions (currently only for 3.9.0) and remove docs mentions of the removed 3.7.0 version.